To streamline the creation of dependent styles, the script that generates the styles has been modified to handle **all** the publishers at once.
The subdirectories in the `data` folder contains the same data as
before, but slightly reformatted to be compatible with the new script.
The format of the info necessary to generate dependent styles is now
more consistent, to make it easier to add more series of journals in
the future. In particular, the placeholders in the templates use the
same name as the column header in the tab-delimited file with the list
of journals, and the relationship is automatically applied (only the
'Title' column is mandatory).  The templates have also been modified to
take advantage of the CSL 1.0.1 specifications.